Move /horizon/internal/codes to /protocols/horizon/codes
