+++ support/txsub/errors.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/stellar/go/services/horizon/internal/codes"
+	"github.com/stellar/go/protocols/horizon/codes"
 	"github.com/stellar/go/xdr"
 )
